feat(web-app): Adjust glucose quality by context and split utilities
- Split `utils/measurements.ts` into `utils/bp.ts` and `utils/glucose.ts`
- Updated all references to use the split modules.
- Adjusted `getGlucoseQuality` logic to bump quality thresholds when `context` is 'post-meal', moving the normal bound from <140 mg/dL up to <180 mg/dL based on ADA guidelines.
