+++ packages/web-app/.env.test
@@ -1,0 +1,1 @@
+VITE_FIREBASE_API_KEY="test_api_key"
