+++ packages/web-app/app/utils/bp.ts
@@ -1,5 +1,4 @@
 import type { BpMeasurement } from '#app/models/measurements/bp.ts';
-import type { GlucoseMeasurement } from '#app/models/measurements/glucose.ts';
 
 export type BpQuality =
   | 'hypertension-crisis'
@@ -8,8 +7,6 @@ export type BpQuality =
   | 'elevated'
   | 'normal'
   | 'low';
-
-export type GlucoseQuality = 'high' | 'elevated' | 'normal' | 'low';
 
 export const getBpQuality = (
   bp: Pick<BpMeasurement, 'systolic' | 'diastolic'>
@@ -31,20 +28,6 @@ export const getBpQuality = (
     return 'low';
   }
   return 'normal';
-};
-
-export const getGlucoseQuality = (
-  glucose: Pick<GlucoseMeasurement, 'value'>
-): GlucoseQuality => {
-  if (glucose.value > 180) {
-    return 'high';
-  } else if (glucose.value > 140) {
-    return 'elevated';
-  } else if (glucose.value >= 80) {
-    return 'normal';
-  } else {
-    return 'low';
-  }
 };
 
 export const BP_STATUS_CLASSES: Record<BpQuality, string> = {
@@ -81,43 +64,4 @@ export const BP_COLORS = {
     bg: 'var(--color-success)',
     fg: 'var(--color-success-content)',
   },
-};
-
-export const GLUCOSE_COLORS = {
-  high: {
-    bg: 'var(--color-error)',
-    fg: 'var(--color-error-content)',
-  },
-  elevated: {
-    bg: 'var(--color-warning)',
-    fg: 'var(--color-warning-content)',
-  },
-  low: {
-    bg: 'var(--color-info)',
-    fg: 'var(--color-info-content)',
-  },
-  normal: {
-    bg: 'var(--color-success)',
-    fg: 'var(--color-success-content)',
-  },
-};
-
-export const GLUCOSE_STATUS_CLASSES: Record<GlucoseQuality, string> = {
-  high: 'status-error',
-  elevated: 'status-warning',
-  normal: 'status-success',
-  low: 'status-info',
-};
-
-const GLUCOSE_CONTEXT = {
-  fasting: { name: 'Fasting' },
-  'post-meal': { name: 'Post-Meal' },
-  other: { name: 'Other' },
-};
-
-export const getGlucoseContextName = (context?: string) => {
-  if (!context) {
-    return 'Unknown';
-  }
-  return GLUCOSE_CONTEXT[context as keyof typeof GLUCOSE_CONTEXT].name;
 };
